Re-done integration tests; fixed bad method name
An friend brought to my attention that pages weren't loading properly.
The reason was that, for whatever reason, I named the view-like method
on the Page model get_rendered_content. As it turns out, in addition to
being misleading, it's actually running over the method of the same
name (with different purpose) on the ContentArea class it inherits
from.

Sadly, the tests didn't catch this because they were only checking for
the title in the content, and I suppose the error was being suppressed,
rather than a 500 being triggered from the template.

So, I rewrote the test (and the others like it) to be integration ones,
using Client rather than RequestFactory (which doesn't go through
middleware). The test did in fact fail before I renamed the method in
question.